refactor(future-cli): extracted inputFormatHandlers, for reuse
