+++ future-cli.js
@@ -55,6 +55,7 @@ const Blob = require(blobPath).Blob
 
 const OpenJsCad = require(openjscadPath).OpenJsCad
 const openscad = require(openscadPath)
+const makeInputFormatHandlers = require('./js/utils/inputFormatHandlers')
 
 const meta = {
   producer: `OpenJSCAD ${OpenJsCad.version}`,
@@ -83,7 +84,7 @@ const args = process.argv.splice(2)
 // handle arguments
 // inputs, outputs
 let {inputFile, inputFormat, outputFile, outputFormat, gMainParam} = parseArgs(args)
-const inputFormatHandlers = makeInputFormatHandlers()
+const inputFormatHandlers = makeInputFormatHandlers(OpenJsCad, openscad, lib)
 
 // outputs
 const output = determineOutputNameAndFormat(outputFormat, outputFile)
@@ -102,62 +103,6 @@ const outputData = generateOutputData(outputFormat, src, gMainParam)
 writeOutputDataToFile(outputFile, outputData)
 
 // -- helper functions ---------------------------------------------------------------------------------------
-// -- input format handlers
-function makeInputFormatHandlers () {
-  function inputScadHandler (src, inputFile, outputFile) {
-    var scadParser = require('openscad-openjscad-translator') // hardcoded is bad, but works
-    src = scadParser.parse(src) //    doing the magick
-    src = '// producer: OpenJSCAD ' + OpenJsCad.version + '\n' + src
-    src = '// source: ' + outputFile + '\n\n' + src
-    return src
-  }
-
-  function inputStlHandler (src, inputFile) {
-    src = openscad.parseSTL(src, inputFile)
-    return src
-  }
-
-  function inputAmfHandler (src, inputFile) {
-    OpenJsCad = require(path.resolve(lib, './js/jscad-parseAMF.js')).OpenJsCad
-    src = OpenJsCad.parseAMF(src, inputFile)
-    return src
-  }
-
-  function inputObjHandler (src, inputFile) {
-    src = openscad.parseOBJ(src, inputFile)
-    return src
-  }
-
-  function inputGcodeHandler (src, inputFile) {
-    src = openscad.parseGCode(src, inputFile)
-    return src
-  }
-
-  function inputSvgHandler (src, inputFile) {
-    OpenJsCad = require(path.resolve(lib, './js/jscad-parseSVG.js')).OpenJsCad
-    src = OpenJsCad.parseSVG(src, inputFile)
-    return src
-  }
-
-  function inputJsonHandler (src, inputFile) {
-    OpenJsCad = require(path.resolve(lib, './js/jscad-parseJSON.js')).OpenJsCad
-    src = OpenJsCad.parseJSON(src, inputFile)
-    return src
-  }
-
-  const inputFormatHandlers = {
-    'scad': inputScadHandler,
-    'stl': inputStlHandler,
-    'amf': inputAmfHandler,
-    'gcode': inputGcodeHandler,
-    'svg': inputSvgHandler,
-    'json': inputJsonHandler,
-    'jscad': x => x,
-    undefined: x => x
-  }
-  return inputFormatHandlers
-}
-
 function parseArgs (args) {
   // hint: https://github.com/substack/node-optimist
   //       https://github.com/visionmedia/commander.js
